refactor: Moved model classes to their component folders
