+++ src/app/modules/content/knowledge-node/knowledge-node.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { KnowledgeNode } from '../../../model/knowledge-node/knowledge-node.model';
+import { KnowledgeNode } from './model/knowledge-node.model';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { LectureService } from '../lecture/lecture.service';
