+++ src/app/modules/content/knowledge-node/model/knowledge-node.model.spec.ts
