+++ src/app/modules/content/knowledge-node/model/knowledge-type.enum.ts
