+++ src/app/modules/content/learning-objects/arrange-task/model/arrange-task.model.ts
@@ -1,7 +1,7 @@
-import { LearningObject } from '../learning-object/learning-object.model';
-import { LearningObjectComponent } from '../../modules/content/learning-objects/learning-object-component';
+import { LearningObject } from '../../model/learning-object.model';
+import { LearningObjectComponent } from '../../learning-object-component';
 import { Type } from '@angular/core';
-import { ArrangeTaskComponent } from '../../modules/content/learning-objects/arrange-task/arrange-task.component';
+import { ArrangeTaskComponent } from '../arrange-task.component';
 import { Container } from './container.model';
 import { Element } from './element.model';
