+++ src/app/modules/content/learning-objects/challenge/model/challenge.model.spec.ts
