+++ src/app/modules/content/learning-objects/image/model/image.model.spec.ts
