+++ src/app/modules/content/learning-objects/text/model/text.model.ts
@@ -1,6 +1,6 @@
-import { LearningObject } from '../learning-object/learning-object.model';
-import { TextComponent } from '../../modules/content/learning-objects/text/text.component';
-import { LearningObjectComponent } from '../../modules/content/learning-objects/learning-object-component';
+import { LearningObject } from '../../model/learning-object.model';
+import { TextComponent } from '../text.component';
+import { LearningObjectComponent } from '../../learning-object-component';
 import { Type } from '@angular/core';
 
 export class Text extends LearningObject {
