+++ src/app/modules/content/learning-objects/video/video.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { LearningObjectComponent } from '../learning-object-component';
-import { Video } from '../../../../model/video/video.model';
+import { Video } from './model/video.model';
 
 @Component({
   selector: 'cc-video',
