+++ src/app/modules/content/lecture/model/lecture.model.ts
