+++ src/app/modules/notes/model/note.model.ts
